requirements_gpu.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
dill==0.3.1.1
future==0.18.2
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
promise==2.3
prompt-toolkit==3.0.2
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
sklearn==0.0
tb-nightly==2.1.0a20191206
tensorboard==2.0.2
tensorflow-datasets==1.3.2
tensorflow-estimator==2.0.1
tensorflow-estimator-2.0-preview==2.0.0
tensorflow-gpu==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.15.2
termcolor==1.1.0
tf-nightly-gpu-2.0-preview==2.0.0.dev20191002
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
